requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
transformers==4.38.2
torch==2.2.1
streamlit==1.32.0
fastapi==0.110.0
uvicorn==0.29.0
python-dotenv==1.0.1
newspaper3k==0.2.8
scikit-learn==1.4.1.post1
spacy==3.7.4
matplotlib==3.7.1
wordcloud==1.8.2.2
pandas==2.2.1
TTS==0.22.0
loguru==0.7.2
sentencepiece==0.2.0
sacremoses==0.1.0